refactor: extract emart24 common code into emart24_common.py
